split workspace_provider to avoid circular dependencies re #306
