+++ mslice/models/projection/powder/mantid_projection_calculator.py
@@ -1,7 +1,9 @@
 from __future__ import (absolute_import, division, print_function)
 
-from mslice.models.workspacemanager.workspace_provider import get_workspace_handle, propagate_properties, run_algorithm
+from mslice.models.workspacemanager.workspace_algorithms import propagate_properties
+from mslice.models.workspacemanager.workspace_provider import get_workspace_handle
 from mslice.models.projection.powder.projection_calculator import ProjectionCalculator
+from mslice.util.mantid import run_algorithm
 import mantid.simpleapi as s_api
 from ...labels import DELTA_E_LABEL, MEV_LABEL, MOD_Q_LABEL, WAVENUMBER_LABEL, THETA_LABEL
