+++ mslice/models/workspacemanager/file_io.py
@@ -2,7 +2,8 @@ from __future__ import (absolute_import, division, print_function)
 import os.path
 from mantid.api import MDNormalization
 from mslice.util.qt.QtWidgets import QFileDialog
-from mslice.models.workspacemanager.workspace_provider import get_workspace_handle, run_alg
+from mslice.models.workspacemanager.workspace_algorithms import run_alg
+from mslice.models.workspacemanager.workspace_provider import get_workspace_handle
 from mslice.workspace.histogram_workspace import HistogramWorkspace
 
 import numpy as np
